useLocalIdentity: handle remove case (#930)
